requirements.txt
six>=1.10.0
pandas>=0.17.1
seaborn>=0.7.0
click>=6.6
munch>=2.0.4
pyyaml>=3.11
